refactor: Move `HoyoverseGame` interface into interfaces package
