+++ src/main/java/ovh/akio/hmu/Application.java
@@ -4,10 +4,7 @@ import me.tongfei.progressbar.ProgressBar;
 import ovh.akio.hmu.entities.PckAudioFile;
 import ovh.akio.hmu.entities.UpdatePackage;
 import ovh.akio.hmu.entities.WemAudioFile;
-import ovh.akio.hmu.exceptions.InvalidGameDirectoryException;
-import ovh.akio.hmu.games.GenshinImpactGame;
-import ovh.akio.hmu.games.HonkaiStarRail;
-import ovh.akio.hmu.games.HoyoverseGame;
+import ovh.akio.hmu.interfaces.HoyoverseGame;
 import ovh.akio.hmu.interfaces.AudioConverter;
 import ovh.akio.hmu.interfaces.AudioFile;
 import ovh.akio.hmu.wrappers.HDiffPatchWrapper;
